+++ MADCFE/MADCFE/Forms/Empleados/Usuarios.vb
@@ -17,11 +17,21 @@ Public Class Usuarios
         Dim fecha As Date = dtpUsuarios_FechaNacimiento.Value
         Dim gender As String = cbUsuarios_Genero.SelectedItem.ToString()
 
+        If txtUsuarios_Username.Text <> "" And txtUsuarios_Password.Text <> "" And cbUsuarios_TipoUsuario.Text <> "" And
+            txtUsuarios_Nombre.Text <> "" And txtUsuarios_Curp.Text <> "" And txtUsuarios_Rfc.Text <> "" And
+            dtpUsuarios_FechaNacimiento.Value <> Nothing And cbUsuarios_Genero.Text <> "" Then
 
-        Dim qry As New Usuarios_Querys()
-        qry.InsertUsuario(username, password, userType, fullname, curp, rfc, fecha, gender, Globals.username)
-        dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
-        dgvUsuarios.DataMember = "users"
+            Dim qry As New Usuarios_Querys()
+            qry.InsertUsuario(username, password, userType, fullname, curp, rfc, fecha, gender, Globals.username)
+            dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
+            dgvUsuarios.DataMember = "users"
+
+
+        Else
+            MessageBox.Show("Faltan Datos", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error)
+        End If
+
+
     End Sub
 
     Private Sub dgvUsuarios_RowHeaderMouseClick(sender As Object, e As DataGridViewCellMouseEventArgs) Handles dgvUsuarios.RowHeaderMouseClick
@@ -60,18 +70,43 @@ Public Class Usuarios
         Dim blocked As Boolean = chbUsuarios_Bloqueado.Checked
         Dim active As Boolean = chbUsuarios_Activo.Checked
 
-        Dim qry As New Usuarios_Querys()
-        qry.UpdateUsuario(id, username, password, userType, fullname, curp, rfc, fecha, gender, tried, Globals.username, blocked, active)
-        dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
-        dgvUsuarios.DataMember = "users"
+
+        If txtUsuarios_Username.Text <> "" And txtUsuarios_Password.Text <> "" And cbUsuarios_TipoUsuario.Text <> "" And
+            txtUsuarios_Nombre.Text <> "" And txtUsuarios_Curp.Text <> "" And txtUsuarios_Rfc.Text <> "" And
+            dtpUsuarios_FechaNacimiento.Value <> Nothing And cbUsuarios_Genero.Text <> "" Then
+
+            Dim qry As New Usuarios_Querys()
+            qry.UpdateUsuario(id, username, password, userType, fullname, curp, rfc, fecha, gender, tried, Globals.username, blocked, active)
+            dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
+            dgvUsuarios.DataMember = "users"
+
+
+        Else
+            MessageBox.Show("Faltan Datos", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error)
+
+        End If
+
+
+
     End Sub
 
     Private Sub btnUsuarios_Eliminar_Click(sender As Object, e As EventArgs) Handles btnUsuarios_Eliminar.Click
-        Dim id As String = lblUsuarios_id.Text
-        Dim qry As New Usuarios_Querys()
-        qry.DeleteUsuario(id)
-        dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
-        dgvUsuarios.DataMember = "users"
+
+        Dim Respuesta As String = MessageBox.Show("¿Estas seguro de que quieres borrar a este usuario?", "¡CUIDADO!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning)
+
+        If Respuesta = vbYes Then
+            Dim id As String = lblUsuarios_id.Text
+            Dim qry As New Usuarios_Querys()
+            qry.DeleteUsuario(id)
+            dgvUsuarios.DataSource = qry.SelectAllFromUsuarios()
+            dgvUsuarios.DataMember = "users"
+
+            MessageBox.Show("Datos Eliminados con Exito", ":)", MessageBoxButtons.OK, MessageBoxIcon.Information)
+
+        End If
+
+
+
     End Sub
 
     Private Sub txtUsuarios_Nombre_KeyPress(sender As Object, e As KeyPressEventArgs) Handles txtUsuarios_Nombre.KeyPress
